% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{hello}
\alias{hello}
\title{Our first example function}
\usage{
hello(x, ...)
}
\arguments{
\item{x}{number}

\item{...}{anything you like!}
}
\value{
another number
}
\description{
Our first example function
}
\examples{
hello(3)
}
