% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{infect_corner}
\alias{infect_corner}
\title{Initial matrix with corners infected}
\usage{
infect_corner(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\value{
SIRmatrix
}
\description{
Initial matrix with corners infected
}
\examples{
infect_corner(4,6)
}
