% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{plot.SIRmatrix}
\alias{plot.SIRmatrix}
\title{Plot the SIRmatrix}
\usage{
\method{plot}{SIRmatrix}(x, col = c("white", "red", "gray"), ...)
}
\arguments{
\item{x}{SIRMatrix}

\item{col}{list of three colors you want to represent SIR}

\item{...}{any other parameter to pass into image()}
}
\value{
plot of the matrix
}
\description{
Plot the SIRmatrix
}
\examples{
A = random_infect(10,10)
plot(A)

}
