% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{random_infect}
\alias{random_infect}
\title{Randomly infect SIR model.}
\usage{
random_infect(rows, columns, proportion = 0.1)
}
\arguments{
\item{rows}{number of rows}

\item{columns}{number of columns}

\item{proportion}{proportion to initally infect population}
}
\value{
matrix with randomly infected population of proportion p
}
\description{
Randomly infect SIR model.
}
\examples{
random_infect(8,8)
}
