% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{simulate_sir}
\alias{simulate_sir}
\title{Simulating SIRmatrix until no infected cells remain}
\usage{
simulate_sir(A, prob)
}
\arguments{
\item{A}{SIRmatrix}

\item{prob}{probability to infect neighbors}
}
\value{
list of iterations, probability, proportion infected, and final matrix
}
\description{
Simulating SIRmatrix until no infected cells remain
}
\examples{
A = infect_corner(10,10)
simulate_sir(A, 0.25)
}
