% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{step}
\alias{step}
\title{One step of SIRmatrix model}
\usage{
step(x, prob)
}
\arguments{
\item{x}{SIRmatrix}

\item{prob}{probability of infected neighbors of an infected cell}
}
\value{
SIRmatrix
}
\description{
One step of SIRmatrix model
}
\examples{
A = infect_corner(10,10)
A =step(A, 0.25)
}
