% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{summary.SIRmatrix}
\alias{summary.SIRmatrix}
\title{Summary of each column of SIR matrix}
\usage{
\method{summary}{SIRmatrix}(object, ...)
}
\arguments{
\item{object}{matrix}

\item{...}{any other parameters to be passed into summary()}
}
\value{
summary list showing total cells, Susceptible, Infected, Removed
}
\description{
Summary of each column of SIR matrix
}
\examples{
A = infect_corner(10,10)
summary(A)
}
